{{ cookiecutter.project_name }}/pyproject.toml
[build-system]
requires = ["uv_build>=0.6"]
build-backend = "uv_build"

[project]
name = "{{ cookiecutter.project_name }}"
version = "{{ cookiecutter.project_version }}"
description = "{{ cookiecutter.project_short_description }}"

authors = [{ name = "{{ cookiecutter.full_name }}", email = "{{ cookiecutter.email }}" }]
maintainers = [{ name = "{{ cookiecutter.full_name }}", email = "{{ cookiecutter.email }}" }]

requires-python = ">={{ cookiecutter.python_version }}"
readme = "README.md"
license = { file = "LICENSE" }

dependencies = [
    "clu>=0.0.12",
    "jax-ai-stack[grain]>=0.4.9",
    "tf-nightly-cpu>=2.20.0.dev20250214",
]

[project.optional-dependencies]
cuda = ["jax-cuda12-plugin[with-cuda]>=0.6.0"]

[tool.uv.sources]
jax-ai-stack = { git = "https://github.com/jax-ml/jax-ai-stack" }

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.1",
    "mkdocs-material>=9.6.12",
    "penzai>=0.2.5",
    "pre-commit>=4.2.0",
    "ruff>=0.11.6",
]
